Relax restrictions on commenting.
Previously, it wouldn't let the person that submitted the proposal leave
a comment on their own proposal, and wouldn't let a person that just
submitted a comment do so again. Unfortunately, allowing these is useful
for people to leave us comments or leave another comment if they realized
they forgot to mention something.