Update fixture names and classes; set up shortcuts
Fixture files need to be named after the corresponding database tables.
With namespaced models, they're not able to automatically infer what
model class to use, so we configure that manually in the spec_helper.

We also define shortcut methods for each fixture class, so we can
continue to call users(:x), instead of open_conference_ware_users(:x).